Break up Subclasser.
